Organize enforcement domain files
