+++ src/AppServices/Enforcement/EnforcementActionService.cs
@@ -1,6 +1,7 @@
 ﻿using AirWeb.AppServices.Enforcement.EnforcementActionCommand;
 using AirWeb.AppServices.Users;
-using AirWeb.Domain.EnforcementEntities;
+using AirWeb.Domain.EnforcementEntities.CaseFiles;
+using AirWeb.Domain.EnforcementEntities.EnforcementActions;
 
 namespace AirWeb.AppServices.Enforcement;
