+++ src/Domain/EnforcementEntities/EnforcementActions/IEnforcementActionRepository.cs
@@ -1,0 +1,3 @@
+﻿namespace AirWeb.Domain.EnforcementEntities.EnforcementActions;
+
+public interface IEnforcementActionRepository : IRepository<EnforcementAction>;
