--- src/Domain/EnforcementEntities/IEnforcementActionRepository.cs
@@ -1,5 +1,0 @@
-﻿using AirWeb.Domain.EnforcementEntities.EnforcementActions;
-
-namespace AirWeb.Domain.EnforcementEntities;
-
-public interface IEnforcementActionRepository : IRepository<EnforcementAction>;
